feb11/backend/requirements.txt
# Invoice Automation System - Requirements
# Python 3.11+ recommended
# For Windows 11 / Cursor IDE

# ============================================================================
# Core Framework
# ============================================================================
gradio>=5.0.0

# ============================================================================
# FastAPI Backend (NEW)
# ============================================================================
fastapi>=0.104.0          # FastAPI framework
uvicorn[standard]>=0.24.0 # ASGI server with performance extras
python-multipart>=0.0.6   # For file upload support

# ============================================================================
# OpenAI API
# ============================================================================
openai>=1.0.0

# ============================================================================
# Data Processing
# ============================================================================
pandas>=2.0.0
openpyxl>=3.1.0

# ============================================================================
# Database
# ============================================================================
pymongo>=4.0.0

# ============================================================================
# Environment Variables
# ============================================================================
python-dotenv>=1.0.0

# ============================================================================
# PDF/PNG Export
# ============================================================================
Pillow>=10.0.0
reportlab>=4.0.0
weasyprint>=60.0

# ============================================================================
# Browser Automation (for PDF generation)
# ============================================================================
playwright>=1.47.0

# ============================================================================
# Optional: Production Deployment
# ============================================================================
# gunicorn>=21.2.0        # Production WSGI server (Linux/Mac only)
# python-jose[cryptography]>=3.3.0  # For JWT authentication
# passlib[bcrypt]>=1.7.4  # For password hashing

# ============================================================================
# Additional Dependencies (auto-installed by above packages)
# ============================================================================
# numpy>=1.24.0
# requests>=2.31.0
# typing-extensions>=4.0.0


